[cli] better detection of background OT CLI commands + add more background commands
Improves the detection of background OT CLI commands by checking command output of potential OT CLI
background commands for the 'Done' string at the end. If 'Done' is present, the command is marked
as a foreground (non-background) command at that point. This allows a simpler definition of the list
of (potential) background commands and fixes the issue that some commands may not work correctly
when used with the 'async' flag.

Also a unit test is added to exercise various background commands.